added the filter in api gateway fetching id with token

+++ transaction-service/src/main/java/com/splitwise/transaction_service/repository/TransactionDetailRepository.java
@@ -3,5 +3,5 @@ package com.splitwise.transaction_service.repository;
 import com.splitwise.transaction_service.document.TransactionDetail;
 import org.springframework.data.mongodb.repository.MongoRepository;
 
-public interface TransactionDetailRepository extends MongoRepository<TransactionDetail, Integer> {
+public interface TransactionDetailRepository extends MongoRepository<TransactionDetail, Long> {
 }
